Minor fixes - Layout - Translation - View/Modal file split
