+++ app/assets/stylesheets/themes/_theme-template.scss
@@ -235,7 +235,7 @@
         }
     }
     .input-group-unbordered-before {
-        > span > span {
+        > span > span:first-of-type {
             background: none !important;
             border: none !important;
         }
