+++ app/components/Modal/BorrowModal.jsx
@@ -1,363 +1,19 @@
 import React from "react";
 import PropTypes from "prop-types";
 import ZfApi from "react-foundation-apps/src/utils/foundation-api";
-import Translate from "react-translate-component";
 import ChainTypes from "../Utility/ChainTypes";
 import ReactTooltip from "react-tooltip";
 import BindToChainState from "../Utility/BindToChainState";
-import FormattedAsset from "../Utility/FormattedAsset";
 import utils from "common/utils";
-import AmountSelector from "../Utility/AmountSelectorStyleGuide";
-import BalanceComponent from "../Utility/BalanceComponent";
 import WalletApi from "api/WalletApi";
 import WalletDb from "stores/WalletDb";
-import FormattedPrice from "../Utility/FormattedPrice";
 import counterpart from "counterpart";
-import HelpContent from "../Utility/HelpContent";
 import Immutable from "immutable";
 import {ChainStore} from "bitsharesjs";
 import {List} from "immutable";
-import {
-    Checkbox,
-    Modal,
-    Button,
-    Tooltip,
-    Form,
-    Slider,
-    Input,
-    Icon
-} from "bitshares-ui-style-guide";
+import {Modal, Button} from "bitshares-ui-style-guide";
 import asset_utils from "../../lib/common/asset_utils";
-
-const BorrowModalView = props => {
-    const {
-        // Objects
-        accountObj,
-        backingAssetObj,
-        collateralBalanceObj,
-        debtBalanceObj,
-        quoteAssetObj,
-        errors,
-
-        // Strings, Floats and Numbers
-        collateral,
-        collateral_ratio,
-        debtAmount,
-        maintenanceRatio,
-        remainingBackingBalance,
-        target_collateral_ratio,
-        unlockedInputType,
-
-        // Bool Flags
-        disableHelp,
-        isRatioLocked,
-        isOriginalBelowMCR,
-        isPredictionMarket,
-        isValid,
-        useTargetCollateral
-    } = props;
-
-    let quotePrecision = utils.get_asset_precision(
-        quoteAssetObj.get("precision")
-    );
-
-    const userExchangePrice = this.state.newPosition ? (
-        <FormattedPrice
-            noPopOver
-            noTip
-            quote_amount={
-                maintenanceRatio * this.state.debtAmount * quotePrecision
-            }
-            quoteAssetObj={quoteAssetObj.get("id")}
-            base_asset={backingAssetObj.get("id")}
-            base_amount={this.state.collateral * backingPrecision}
-        />
-    ) : null;
-
-    const noValidComponent = (
-        <div style={{textAlign: "center"}}>
-            <Translate
-                component="h3"
-                content="borrow.no_valid"
-                asset_symbol={quoteAssetObj.get("symbol")}
-            />
-        </div>
-    );
-
-    const bitAssetBalanceText = (
-        <span>
-            <span>
-                {collateral != 0 ? (
-                    <span>
-                        <Translate
-                            component="a"
-                            onClick={props.onPayDebt.bind(this)}
-                            content="borrow.pay_max_debt"
-                        />
-                        &nbsp;
-                    </span>
-                ) : null}
-                <Translate component="span" content="transfer.available" />:{" "}
-                <span>
-                    {debtBalanceObj.id ? (
-                        <BalanceComponent balance={debtBalanceObj.id} />
-                    ) : (
-                        <FormattedAsset
-                            amount={0}
-                            asset={quoteAssetObj.get("id")}
-                        />
-                    )}
-                </span>
-            </span>
-        </span>
-    );
-
-    const backingBalanceText = (
-        <span>
-            <span>
-                <span>
-                    <Translate
-                        component="a"
-                        onClick={props.onMaximizeCollatereal.bind(this)}
-                        content="borrow.use_max"
-                    />
-                    &nbsp;
-                </span>
-                <Translate component="span" content="transfer.available" />:{" "}
-                <span>
-                    {collateralBalanceObj.id ? (
-                        <FormattedAsset
-                            amount={remainingBackingBalance}
-                            asset={backingAssetObj.get("id")}
-                        />
-                    ) : (
-                        <FormattedAsset
-                            amount={0}
-                            asset={backingAssetObj.get("id")}
-                        />
-                    )}
-                </span>
-            </span>
-        </span>
-    );
-
-    return !isValid ? (
-        noValidComponent
-    ) : (
-        <div style={{textAlign: "left"}}>
-            {disableHelp ? null : (
-                <HelpContent
-                    path={
-                        "components/" +
-                        (isPredictionMarket
-                            ? "BorrowModalPrediction"
-                            : "BorrowModal")
-                    }
-                    debt={quoteAssetObj.get("symbol")}
-                    collateral={backingAssetObj.get("symbol")}
-                    borrower={accountObj.get("name")}
-                    mr={maintenanceRatio}
-                />
-            )}
-
-            {isOriginalBelowMCR ? (
-                <Translate
-                    component="h6"
-                    className="has-warning"
-                    content="borrow.errors.below_info"
-                />
-            ) : null}
-
-            {!isPredictionMarket ? (
-                <div
-                    style={{
-                        paddingTop: "1rem",
-                        paddingBottom: "1rem"
-                    }}
-                >
-                    <div className="borrow-price-feeds">
-                        <span className="borrow-price-label">
-                            <Translate content="transaction.feed_price" />
-                            :&nbsp;
-                        </span>
-                        <FormattedPrice
-                            noPopOver
-                            quote_amount={asset_utils
-                                .extractRawFeedPrice(quoteAssetObj)
-                                .getIn(["base", "amount"])}
-                            quoteAssetObj={asset_utils
-                                .extractRawFeedPrice(quoteAssetObj)
-                                .getIn(["base", "asset_id"])}
-                            base_asset={asset_utils
-                                .extractRawFeedPrice(quoteAssetObj)
-                                .getIn(["quote", "asset_id"])}
-                            base_amount={asset_utils
-                                .extractRawFeedPrice(quoteAssetObj)
-                                .getIn(["quote", "amount"])}
-                        />
-                    </div>
-                    <b />
-                    <div
-                        className={
-                            "borrow-price-final " +
-                            (errors.below_maintenance
-                                ? "has-error"
-                                : errors.close_maintenance
-                                    ? "has-warning"
-                                    : "")
-                        }
-                    >
-                        <span className="borrow-price-label">
-                            <Translate content="exchange.your_price" />
-                            :&nbsp;
-                        </span>
-                        {userExchangePrice}
-                    </div>
-                </div>
-            ) : null}
-
-            <Form className="full-width" layout="vertical">
-                <AmountSelector
-                    label="transaction.borrow_amount"
-                    amount={debtAmount.toString()}
-                    onChange={props.onBorrowChange.bind(this)}
-                    asset={quoteAssetObj.get("id")}
-                    assets={[quoteAssetObj.get("id")]}
-                    display_balance={bitAssetBalanceText}
-                    placeholder="0.0"
-                    tabIndex={1}
-                    lockStatus={unlockedInputType == "debt" ? false : true}
-                    onLockChange={props.onLockChangeDebt.bind(this)}
-                />
-                <AmountSelector
-                    label="transaction.collateral"
-                    amount={collateral.toString()}
-                    onChange={props.onCollateralChange.bind(this)}
-                    asset={backingAssetObj.get("id")}
-                    assets={[backingAssetObj.get("id")]}
-                    display_balance={backingBalanceText}
-                    placeholder="0.0"
-                    tabIndex={2}
-                    lockStatus={
-                        unlockedInputType == "collateral" ? false : true
-                    }
-                    onLockChange={props.onLockChangeCollateral.bind(this)}
-                    validateStatus={errors.collateral_balance ? "error" : ""}
-                    help={
-                        errors.collateral_balance
-                            ? errors.collateral_balance
-                            : null
-                    }
-                />
-                {!isPredictionMarket ? (
-                    <Form.Item
-                        label={counterpart.translate("borrow.coll_ratio")}
-                        validateStatus={
-                            errors.close_maintenance
-                                ? "warning"
-                                : errors.below_maintenance
-                                    ? "error"
-                                    : null
-                        }
-                        help={
-                            errors.close_maintenance
-                                ? errors.close_maintenance
-                                : errors.below_maintenance
-                                    ? errors.below_maintenance
-                                    : null
-                        }
-                    >
-                        <Input.Group compact>
-                            <Input
-                                style={{width: "30%"}}
-                                value={
-                                    collateral_ratio == 0
-                                        ? ""
-                                        : collateral_ratio.toFixed(2)
-                                }
-                                tabIndex={3}
-                                onChange={props.onRatioChange.bind(this)}
-                                className="input-group-unbordered-before"
-                                addonBefore={
-                                    <Tooltip
-                                        title="Locking Collateral Ratio will override other locked fields"
-                                        placement="right"
-                                    >
-                                        <Icon
-                                            className={
-                                                !isRatioLocked
-                                                    ? "green"
-                                                    : "blue"
-                                            }
-                                            type={
-                                                !isRatioLocked
-                                                    ? "unlock"
-                                                    : "lock"
-                                            }
-                                            onClick={props.onLockChangeCR.bind(
-                                                this
-                                            )}
-                                            style={{fontSize: "20px"}}
-                                        />
-                                    </Tooltip>
-                                }
-                            />
-                            <Slider
-                                style={{width: "67%", marginTop: 10}}
-                                step={0.01}
-                                min={0}
-                                max={maintenanceRatio * 6}
-                                value={collateral_ratio}
-                                onChange={props.onRatioChange.bind(this)}
-                            />
-                        </Input.Group>
-                    </Form.Item>
-                ) : null}
-                {!isPredictionMarket ? (
-                    <Form.Item
-                        validateStatus={
-                            errors.tcr_below_maintenance ? "error" : ""
-                        }
-                        help={
-                            errors.tcr_below_maintenance
-                                ? errors.tcr_below_maintenance
-                                : null
-                        }
-                    >
-                        <Input.Group compact>
-                            <Checkbox
-                                onClick={props.onSetUseTCR.bind(this)}
-                                checked={useTargetCollateral}
-                                tabIndex={4}
-                            >
-                                <Translate content="borrow.enable_target_collateral_ratio" />
-                            </Checkbox>
-                            <Tooltip
-                                title={counterpart.translate(
-                                    "tooltip.target_collateral_ratio"
-                                )}
-                            >
-                                <Icon type="question-circle" />
-                            </Tooltip>
-                        </Input.Group>
-
-                        {useTargetCollateral ? (
-                            <Input
-                                value={
-                                    isNaN(target_collateral_ratio)
-                                        ? "0"
-                                        : target_collateral_ratio
-                                }
-                                tabIndex={5}
-                                onChange={props.onTCRatioChange.bind(this)}
-                            />
-                        ) : null}
-                    </Form.Item>
-                ) : null}
-            </Form>
-        </div>
-    );
-};
+import {BorrowModalView} from "./BorrowModalView";
 
 /**
  *  Given an account and an asset id, render a modal allowing modification of a margin position for that asset
@@ -416,7 +72,7 @@ class BorrowModalContent extends React.Component {
                     collateral: collateral,
                     target_collateral_ratio: target_collateral_ratio
                 },
-                unlockedInputType: "debt",
+                unlockedInputType: "collateral",
                 isRatioLocked: true
             };
         } else {
@@ -1073,11 +729,13 @@ class BorrowModalContent extends React.Component {
                     collateralBalanceObj={collateralBalanceObj}
                     debtBalanceObj={debtBalanceObj}
                     quoteAssetObj={quoteAssetObj}
+                    newPosition={this.state.newPosition || null}
                     errors={errors}
                     // Strings, Floats and Numbers
                     collateral={collateral}
                     collateral_ratio={collateral_ratio}
                     debtAmount={debtAmount}
+                    backingPrecision={backingPrecision}
                     maintenanceRatio={maintenanceRatio}
                     remainingBackingBalance={remainingBackingBalance}
                     target_collateral_ratio={target_collateral_ratio}
